+++ src/rpc/handlers/Unsubscribe.hpp
@@ -21,14 +21,8 @@
 
 #include "data/BackendInterface.hpp"
 #include "feed/SubscriptionManagerInterface.hpp"
-#include "rpc/Errors.hpp"
-#include "rpc/JS.hpp"
-#include "rpc/RPCHelpers.hpp"
-#include "rpc/common/Checkers.hpp"
 #include "rpc/common/Specs.hpp"
 #include "rpc/common/Types.hpp"
-#include "rpc/common/Validators.hpp"
-#include "util/Assert.hpp"
 
 #include <boost/json/conversion.hpp>
 #include <boost/json/value.hpp>
@@ -41,8 +35,6 @@
 #include <memory>
 #include <optional>
 #include <string>
-#include <string_view>
-#include <variant>
 #include <vector>
 
 namespace feed {
@@ -94,10 +86,7 @@ public:
     UnsubscribeHandler(
         std::shared_ptr<BackendInterface> const& sharedPtrBackend,
         std::shared_ptr<feed::SubscriptionManagerInterface> const& subscriptions
-    )
-        : sharedPtrBackend_(sharedPtrBackend), subscriptions_(subscriptions)
-    {
-    }
+    );
 
     /**
      * @brief Returns the API specification for the command
@@ -106,40 +95,7 @@ public:
      * @return The spec for the given apiVersion
      */
     static RpcSpecConstRef
-    spec([[maybe_unused]] uint32_t apiVersion)
-    {
-        static auto const booksValidator =
-            validation::CustomValidator{[](boost::json::value const& value, std::string_view key) -> MaybeError {
-                if (!value.is_array())
-                    return Error{Status{RippledError::rpcINVALID_PARAMS, std::string(key) + "NotArray"}};
-
-                for (auto const& book : value.as_array()) {
-                    if (!book.is_object())
-                        return Error{Status{RippledError::rpcINVALID_PARAMS, std::string(key) + "ItemNotObject"}};
-
-                    if (book.as_object().contains("both") && !book.as_object().at("both").is_bool())
-                        return Error{Status{RippledError::rpcINVALID_PARAMS, "bothNotBool"}};
-
-                    auto const parsedBook = parseBook(book.as_object());
-                    if (auto const status = std::get_if<Status>(&parsedBook))
-                        return Error(*status);
-                }
-
-                return MaybeError{};
-            }};
-
-        static auto const rpcSpec = RpcSpec{
-            {JS(streams), validation::SubscribeStreamValidator},
-            {JS(accounts), validation::SubscribeAccountsValidator},
-            {JS(accounts_proposed), validation::SubscribeAccountsValidator},
-            {JS(books), booksValidator},
-            {JS(url), check::Deprecated{}},
-            {JS(rt_accounts), check::Deprecated{}},
-            {"rt_transactions", check::Deprecated{}},
-        };
-
-        return rpcSpec;
-    }
+    spec([[maybe_unused]] uint32_t apiVersion);
 
     /**
      * @brief Process the Unsubscribe command
@@ -149,119 +105,35 @@ public:
      * @return The result of the operation
      */
     Result
-    process(Input input, Context const& ctx) const
-    {
-        if (input.streams)
-            unsubscribeFromStreams(*(input.streams), ctx.session);
-
-        if (input.accounts)
-            unsubscribeFromAccounts(*(input.accounts), ctx.session);
-
-        if (input.accountsProposed)
-            unsubscribeFromProposedAccounts(*(input.accountsProposed), ctx.session);
-
-        if (input.books)
-            unsubscribeFromBooks(*(input.books), ctx.session);
-
-        return Output{};
-    }
+    process(Input input, Context const& ctx) const;
 
 private:
     void
     unsubscribeFromStreams(std::vector<std::string> const& streams, std::shared_ptr<web::ConnectionBase> const& session)
-        const
-    {
-        for (auto const& stream : streams) {
-            if (stream == "ledger") {
-                subscriptions_->unsubLedger(session);
-            } else if (stream == "transactions") {
-                subscriptions_->unsubTransactions(session);
-            } else if (stream == "transactions_proposed") {
-                subscriptions_->unsubProposedTransactions(session);
-            } else if (stream == "validations") {
-                subscriptions_->unsubValidation(session);
-            } else if (stream == "manifests") {
-                subscriptions_->unsubManifest(session);
-            } else if (stream == "book_changes") {
-                subscriptions_->unsubBookChanges(session);
-            } else {
-                ASSERT(false, "Unknown stream: {}", stream);
-            }
-        }
-    }
+        const;
 
     void
     unsubscribeFromAccounts(std::vector<std::string> accounts, std::shared_ptr<web::ConnectionBase> const& session)
-        const
-    {
-        for (auto const& account : accounts) {
-            auto const accountID = accountFromStringStrict(account);
-            subscriptions_->unsubAccount(*accountID, session);
-        }
-    }
+        const;
 
     void
     unsubscribeFromProposedAccounts(
         std::vector<std::string> accountsProposed,
         std::shared_ptr<web::ConnectionBase> const& session
-    ) const
-    {
-        for (auto const& account : accountsProposed) {
-            auto const accountID = accountFromStringStrict(account);
-            subscriptions_->unsubProposedAccount(*accountID, session);
-        }
-    }
+    ) const;
 
     void
-    unsubscribeFromBooks(std::vector<OrderBook> const& books, std::shared_ptr<web::ConnectionBase> const& session) const
-    {
-        for (auto const& orderBook : books) {
-            subscriptions_->unsubBook(orderBook.book, session);
+    unsubscribeFromBooks(std::vector<OrderBook> const& books, std::shared_ptr<web::ConnectionBase> const& session)
+        const;
 
-            if (orderBook.both)
-                subscriptions_->unsubBook(ripple::reversed(orderBook.book), session);
-        }
-    }
-
+    /**
+     * @brief Convert a JSON object to an Input
+     *
+     * @param jv The JSON object to convert
+     * @return The Input object
+     */
     friend Input
-    tag_invoke(boost::json::value_to_tag<Input>, boost::json::value const& jv)
-    {
-        auto input = Input{};
-        auto const& jsonObject = jv.as_object();
-
-        if (auto const& streams = jsonObject.find(JS(streams)); streams != jsonObject.end()) {
-            input.streams = std::vector<std::string>();
-            for (auto const& stream : streams->value().as_array())
-                input.streams->push_back(boost::json::value_to<std::string>(stream));
-        }
-        if (auto const& accounts = jsonObject.find(JS(accounts)); accounts != jsonObject.end()) {
-            input.accounts = std::vector<std::string>();
-            for (auto const& account : accounts->value().as_array())
-                input.accounts->push_back(boost::json::value_to<std::string>(account));
-        }
-        if (auto const& accountsProposed = jsonObject.find(JS(accounts_proposed));
-            accountsProposed != jsonObject.end()) {
-            input.accountsProposed = std::vector<std::string>();
-            for (auto const& account : accountsProposed->value().as_array())
-                input.accountsProposed->push_back(boost::json::value_to<std::string>(account));
-        }
-        if (auto const& books = jsonObject.find(JS(books)); books != jsonObject.end()) {
-            input.books = std::vector<OrderBook>();
-            for (auto const& book : books->value().as_array()) {
-                auto internalBook = OrderBook{};
-                auto const& bookObject = book.as_object();
-
-                if (auto const& both = bookObject.find(JS(both)); both != bookObject.end())
-                    internalBook.both = both->value().as_bool();
-
-                auto const parsedBookMaybe = parseBook(book.as_object());
-                internalBook.book = std::get<ripple::Book>(parsedBookMaybe);
-                input.books->push_back(internalBook);
-            }
-        }
-
-        return input;
-    }
+    tag_invoke(boost::json::value_to_tag<Input>, boost::json::value const& jv);
 };
 
 }  // namespace rpc
